# Http[Request|Response]Stream.ValidateState made inlineable