--- a/src/Kestrel.Core/Internal/Http/HttpRequestStream.cs
+++ b/src/Kestrel.Core/Internal/Http/HttpRequestStream.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http.Features;
 using Microsoft.AspNetCore.Connections;
 using Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Infrastructure;
+using System.Runtime.CompilerServices;
 
 namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
 {
@@ -189,29 +190,35 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
             }
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private void ValidateState(CancellationToken cancellationToken)
         {
-            switch (_state)
+            var state = _state;
+            if (state == HttpStreamState.Open)
             {
-                case HttpStreamState.Open:
-                    if (cancellationToken.IsCancellationRequested)
-                    {
-                        cancellationToken.ThrowIfCancellationRequested();
-                    }
-                    break;
-                case HttpStreamState.Closed:
-                    throw new ObjectDisposedException(nameof(HttpRequestStream));
-                case HttpStreamState.Aborted:
-                    if (_error != null)
-                    {
-                        ExceptionDispatchInfo.Capture(_error).Throw();
-                    }
-                    else
-                    {
-                        throw new TaskCanceledException();
-                    }
-                    break;
+                if (cancellationToken.IsCancellationRequested)
+                {
+                    cancellationToken.ThrowIfCancellationRequested();
+                }
             }
+            else if (state == HttpStreamState.Closed)
+            {
+                ThrowObjectDisposedException();
+            }
+            else
+            {
+                if (_error != null)
+                {
+                    ExceptionDispatchInfo.Capture(_error).Throw();
+                }
+                else
+                {
+                    ThrowTaskCanceledException();
+                }
+            }
+
+            void ThrowObjectDisposedException() => throw new ObjectDisposedException(nameof(HttpRequestStream));
+            void ThrowTaskCanceledException() => throw new TaskCanceledException();
         }
     }
 }