Instead of the `switch` `if` is used and the code got tightened.
Now it is inlined in the methods, that other than the call to `ValidateState` just forward to the `Read/Write`-methods.