@gfoidl 2018-05-16
> What motivated this change?

Stepped with the debugger through the source, spotted these methods and I knew that `switch` won't inline. Hence I submitted this PR.

> Any before-after results?

I'm just getting into the Kestrel-source, so can you give me any pointers to relevant benchmarks? Rolling out my own that cover these aspects may be not so easy (I believe).