@halter73 2018-05-16
I don't think we have any microbenchmarks for `Http[Request|Response]Stream` yet. Feel free to add one. For your other PR, there's the [StringUtilitesBenchmark](https://github.com/aspnet/KestrelHttpServer/blob/27799b5acb59523889a3a52a39d606e795b81c50/benchmarks/Kestrel.Performance/StringUtilitiesBenchmark.cs).