@pakrym 2018-05-16
You can also try https://github.com/aspnet/KestrelHttpServer/blob/dev/benchmarks/Kestrel.Performance/InMemoryTransportBenchmark.cs that runs entire pipeline in memory and see if it shows any change.