@halter73 2018-05-16
InMemoryTransportBenchmark is a good idea. I don't think that tests the Http*Request*Stream though.