@gfoidl 2018-05-17
### InMemoryTransportBenchmark

Note I ran the benchmarks with `netcoreapp2.1`, not with `netcoreapp2.0` as in the `csproj`.

#### Current dev

|             Method |     Mean |     Error |    StdDev |      Op/s |  Gen 0 | Allocated |
|------------------- |---------:|----------:|----------:|----------:|-------:|----------:|
|          Plaintext | 7.886 us | 0.1449 us | 0.1285 us | 126,807.0 |      - |     371 B |
| PlaintextPipelined | 3.914 us | 0.0775 us | 0.1060 us | 255,511.0 | 0.1221 |     432 B |

#### This PR

|             Method |     Mean |     Error |    StdDev |   Median |      Op/s |  Gen 0 | Allocated |
|------------------- |---------:|----------:|----------:|---------:|----------:|-------:|----------:|
|          Plaintext | 7.137 us | 0.1814 us | 0.5292 us | 6.869 us | 140,111.2 |      - |     368 B |
| PlaintextPipelined | 3.794 us | 0.0711 us | 0.0665 us | 3.781 us | 263,607.0 | 0.1221 |     432 B |

So it got a bit faster.
`HttpRequestStream` isn't touched by this benchmark. I assume this will also be faster, but not in the same amount as `HttpResponseStream`, due to the more complex validation.