@JunTaoLuo 2018-10-11
I just re-ran the benchmarks on netcoreapp2.2 and here are the results:

Before:
```
             Method |     Mean |     Error |    StdDev |      Op/s | Allocated |
------------------- |---------:|----------:|----------:|----------:|----------:|
          Plaintext | 7.962 us | 0.1587 us | 0.3980 us | 125,598.0 |     369 B |
 PlaintextPipelined | 5.247 us | 0.0906 us | 0.0757 us | 190,575.9 |     432 B |
```
After:
```
             Method |     Mean |     Error |    StdDev |      Op/s | Allocated |
------------------- |---------:|----------:|----------:|----------:|----------:|
          Plaintext | 7.738 us | 0.1529 us | 0.2425 us | 129,240.2 |     369 B |
 PlaintextPipelined | 4.784 us | 0.0421 us | 0.0329 us | 209,020.8 |     432 B |
```

The improvements are modest. I'll merge to release/2.2 and master.