@JunTaoLuo 2018-10-11
Merged in https://github.com/aspnet/KestrelHttpServer/commit/773aa6e6ffa542b30165db4cde3ef3a167f28ffc. Thansk @gfoidl !